feat(app): migrate DynamicSignInPage to SignInPageBlueprint extension
Step 3d.

Move \`DynamicSignInPage\` (the OIDC vs guest auth switcher) out of App.tsx
into its own \`packages/app/src/components/DynamicSignInPage.tsx\` and
register it through \`SignInPageBlueprint.make({ loader: () => import(...) })\`
inside the renamed \`customAppModule\` (was \`customTranslationsModule\`,
now holds both the sign-in extension and the translation overrides since
both target \`pluginId: 'app'\`).

Drop \`components.SignInPage: DynamicSignInPage\` from
\`convertLegacyAppOptions\` and the now-unused
\`SignInPage\`/\`configApiRef\`/\`useApi\`/\`openChoreoAuthApiRef\` imports from
App.tsx.

Verified live: signed out, the OpenChoreo \"Sign in using OpenChoreo\"
button appears via the NFS blueprint, OIDC redirect to Thunder works,
sign-in completes, home page renders.
